// finance_client/src/app/user/data/types/customer-types/item.ts
export interface Item {
  categoryName?: string;
  category_id?: string;
  className?: string;
  commodityCode?: string;
  currency?: string;
  deletedBy?: string;
  deletedFlag?: string;
  deletedTime?: Date;
  description?: string;
  execiseTaxEffectiveStartDate?: Date;
  execiseTaxRate?: number;
  familyName?: string;
  id?: number;
  includeExeciseTaxInformation?: string;
  includePieceOfMeasurementInformation?: string;
  includeVatInformation?: string;
  isService?: string;
  modifiedBy?: string;
  modifiedFlag?: string;
  modifiedTime?: Date;
  name?: string;
  packageMeasurementUnit?: string;
  packageUntiPrice?: number;
  pieceMeasurementUnit?: string;
  pieceQuantity?: number;
  postedFlag?: string;
  postedTime?: Date;
  pricePerPieceQuantity?: number;
  reason?: string;
  segmentName?: string;
  status?: string;
  stock?: number;
  unitOfMeasure?: string;
  unitsPerPieceQuantity?: number;
  vatEffectiveStartDate?: Date;
  vatTaxRate?: number;
  verifiedBy?: string;
  verifiedFlag?: string;
  verifiedTime?: Date;
}